requirements.txt
streamlit==1.54.0
pydantic==2.12.5
sqlalchemy>=2.0
openai==2.21.0
python-dotenv==1.2.1
pinecone==8.0.1
sentence-transformers==5.2.3
pytest==9.0.2